Refactor architecture with Elixir's Behaviours
Use mocks, The Elixir Way™.

We are not mocking something out of nothing. We base our mock upon a
behaviour, to provide another implementation in our tests.
Doing so, we can control the context for the domain to perform actions.

And, in production, we fallback on actual implementations (e.g. GitHub
PRs reader, System Date giver, etc.)